refactor: TextInput 디렉토리 구조 리팩토링
