+++ src/componenets/common/TextInput/TextInput.style.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const CustomInput = styled.input`
+  width: ${({ width }) => width};
+  height: ${({ height }) => height};
+  border: ${({ borderStyle }) => borderStyle};
+  border-radius: ${({ borderRadius }) => borderRadius};
+  background-color: ${({ theme }) => theme.color.gray1};
+
+  & ::placeholder {
+    color: ${({ theme }) => theme.color.gray1};
+  }
+`;
+
+export default CustomInput;
